deps: make aliases2_lookup static, don't export it
Fixes the following OS X dynamic linker issue with node.js master:

    $ $HOME/src/master/out/Release/node test/test-basic.js
    dyld: lazy symbol binding failed: Symbol not found: _aliases2_lookup
      Referenced from: /Users/bnoordhuis/src/node-iconv/build/Debug/iconv.node
      Expected in: dynamic lookup

    dyld: Symbol not found: _aliases2_lookup
      Referenced from: /Users/bnoordhuis/src/node-iconv/build/Debug/iconv.node
      Expected in: dynamic lookup

    Trace/BPT trap: 5

The symbol is defined in deps/libiconv/lib/iconv.c (it's a function)
but for some reason the symbol is marked as undefined in the shared
object:

    $ nm build/Debug/iconv.node | grep aliases2
                     U _aliases2_lookup

Note that it's undefined both with v0.10 and master so it's anybody's
guess why master triggers the lookup error but v0.10 doesn't.  At any
rate, it's easy to work around by making the function static.