lib: fix type coercion bug
Add a .length property with value 0 to the FLUSH sentinel so that its
use in size calculations doesn't evaluate to NaN.

Passing the result to `Buffer.alloc()` before Node.js v12 coerced it
to zero but v12 rejects it with a RangeError exception.

This commit makes test/test-stream.js pass again with v12 pre-releases.

Fixes: https://github.com/bnoordhuis/node-iconv/issues/197